+++ src/App.tsx
@@ -1,7 +1,7 @@
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
-import { Navbar } from './componets/NavBar';
+import { Navbar } from './components/NavBar';
 import { Navigate, Route, Routes } from 'react-router-dom';
 import { HomePage } from './pages/HomePage';
 import { TabsPage } from './pages/TabsPage';
